+++ CHANGELOG.md
@@ -7,6 +7,13 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+## [2.4.1] - 2026-09-23
+
+### Added
+
+- OAuth's `webhooks:write` scope is available in `OAuthScope` and documented for
+  webhook subscription updates and inactivation.
+
 ## [2.4.0] - 2026-09-21
 
 ### Fixed